Crate pushover

Pushover is Pushover API wrapper with attachment support in Rust 2021 edition.

Structs

Attachment

Notification attachment. Image in most cases.

Notification

Pushover API parameters https://pushover.net/api#messages and attachment.

Response

Pushover API response. https://pushover.net/api#response

Enums

AttachmentError

Attachment error.

HTML

To enable HTML formatting. https://pushover.net/api#html

Monospace

To enable monospace messages. https://pushover.net/api#html

NotificationError

Notification error.

Priority

Messages may be sent with a different priority that affects how the message is presented to the user. https://pushover.net/api#priority

Sound

Users can choose from a number of different default sounds to play when receiving notifications. https://pushover.net/api#sounds

Functions

send_notification

Shorthand function to send notification to Pushover.